MV/med_venv/requirements.txt
flask
tensorflow
torch
torchvision
torchxrayvision
transformers
Pillow
soundfile
SpeechRecognition
gtts
